clean up, move stuff; pass scale as appropriate

+++ telemetry/accl.go
@@ -17,9 +17,9 @@ func (accl *ACCL) Parse(bytes []byte, scale *SCAL) error {
 	}
 
 	// ¯\_(ツ)_/¯
-	accl.X = float64(int16(binary.BigEndian.Uint16(bytes[0:2])) / int16(scale.Values[0]))
-	accl.Y = float64(int16(binary.BigEndian.Uint16(bytes[2:4])) / int16(scale.Values[0]))
-	accl.Z = float64(int16(binary.BigEndian.Uint16(bytes[4:6])) / int16(scale.Values[0]))
+	accl.X = float64(int16(binary.BigEndian.Uint16(bytes[0:2]))) / float64(scale.Values[0])
+	accl.Y = float64(int16(binary.BigEndian.Uint16(bytes[2:4]))) / float64(scale.Values[0])
+	accl.Z = float64(int16(binary.BigEndian.Uint16(bytes[4:6]))) / float64(scale.Values[0])
 
 	return nil
 }
